% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPER.R
\name{findSDG}
\alias{findSDG}
\title{Find the spatial dependent gene (SDG) based on SPER scores}
\usage{
findSDG(
  cell.type,
  score.mat,
  extra.gene.list,
  LRP.data,
  marker.gene,
  ratio.threshold = 1.5,
  expr.fraction = 0.15,
  only.extra = T,
  method.name = "SPER"
)
}
\arguments{
\item{cell.type}{Cell Type Name}

\item{score.mat}{Score matrix (SPER, or correlations)}

\item{extra.gene.list}{The gene list of interest: could be extracellular or ligand gene sets}

\item{LRP.data}{Ligand-receptor pair data; used to find the receptors of found paracrine signals}

\item{marker.gene}{The marker gene matrix with a column 'Type' indicating the cell type information and 'Gene' indicating the gene ID}

\item{ratio.threshold}{Threshold on the score: pairs whose scores larger than the threshold will be kept}

\item{expr.fraction}{Threshold on expression prevalence: pairs whose prevalence smaller than the threshold will be kept}

\item{only.extra}{Boolean value: only consider the candidate genes}

\item{method.name}{Name of the method; default is SPER}
}
\value{
A gene list containing the spatial dependent genes
}
\description{
This function helps find the SDG based on the given threshold in both SPER scores and
expression prevalence.
}
