% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPER.R
\name{findSPERsignals}
\alias{findSPERsignals}
\title{Find the putative paracrine signals based on SPER scores}
\usage{
findSPERsignals(
  target.cell,
  score.mat,
  expr.frac.mat,
  marker.gene = NULL,
  gene.set = NULL,
  LRP.data = NULL,
  score.threshold = 1.5,
  expr.fraction = 0.15
)
}
\arguments{
\item{target.cell}{Target ell Type Name}

\item{score.mat}{Score matrix (SPER, or correlations)}

\item{expr.frac.mat}{Expression prevalence matrix: matrix values from 0 to 1, showing the percentage of cells in that cell type expressing the gene.}

\item{marker.gene}{The marker gene matrix with a column 'Type' indicating the cell type information and 'Gene' indicating the gene ID}

\item{gene.set}{The gene list of interest: could be extracellular or ligand gene sets}

\item{LRP.data}{Ligand-receptor pair data; used to find the receptors of found paracrine signals; should contain at least two columns: 'gene1' for the ligands and 'gene2' for receptors}

\item{score.threshold}{Threshold on the score: pairs whose scores larger than the threshold will be kept}

\item{expr.fraction}{Threshold on expression prevalence: pairs whose prevalence smaller than the threshold will be kept}
}
\value{
A gene list containing the spatial dependent genes; if LRP data provided, return a data frame containing both ligand and receptor information
}
\description{
This function helps find the potential signals based on the given threshold in both SPER scores and
expression prevalence. One can provide the marker (cell-type-specific) gene list and ligand-receptor pair data to
limit the search range and get a more precise result.
}
