% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPER.R
\name{getNeighbor}
\alias{getNeighbor}
\title{Identify desired neighbors and calculate weighted expression levels}
\usage{
getNeighbor(neighbor_dist, round.dist, expr.mat)
}
\arguments{
\item{neighbor_dist}{A numeric input that indicates the distance between desired neighboring pairs}

\item{round.dist}{The rounded distance matrix}

\item{expr.mat}{The expression matrix}
}
\value{
A matrix
}
\description{
This function calculates the weighted expression level of a gene to spots lying at a given distance.
}
