% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPER.R
\name{getSPER}
\alias{getSPER}
\title{Calculate the SPER for a distance}
\usage{
getSPER(round.dist, expr.mat, dist.list)
}
\arguments{
\item{round.dist}{The rounded distance matrix}

\item{expr.mat}{The expression matrix}

\item{dist.list}{A vector containing the list of desired distance values}
}
\value{
A list of matrix
}
\description{
This function calculates the weighted expression level of a gene to spots lying at a given distance.
}
