% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPER.R
\name{getSPERmat}
\alias{getSPERmat}
\title{Calculte a part of SPER: spatial weighted expression level}
\usage{
getSPERmat(round.dist, ST.mat, dist.list)
}
\arguments{
\item{round.dist}{The rounded distance matrix}

\item{ST.mat}{The ST matrix}

\item{dist.list}{A vector containing the list of desired distance values}
}
\value{
A list of matrix
}
\description{
This function calculates the weighted expression level of a gene to spots lying at a given distance.
}
