% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{gg_gene}
\alias{gg_gene}
\title{Plot the spatial figures for genes / cell types}
\usage{
gg_gene(feature, spatial_df, coord, title.anno = "Estimated")
}
\arguments{
\item{feature}{Specific spatial feature to plot}

\item{spatial_df}{A data.frame with rows as spatial spots and columns as each features}

\item{coord}{Spatial coordinates that match the spatial data}

\item{title.anno}{Annotation of the title}
}
\value{
a ggplot2 object
}
\description{
With spatial data, coordinates and given specific feature (gene or cell type), this function plots the
map of this feature's spatial distribution.
}
