% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{normalizd}
\alias{normalizd}
\title{Normalize a vector: L2 norm}
\usage{
normalizd(x)
}
\arguments{
\item{x}{An Input vector}
}
\value{
Normalized vector
}
\description{
This function normalizes a given vector so that the length (L2) of a vector is 1
}
