% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotSDG}
\alias{plotSDG}
\title{Plot the spatial dependent gene (SDG) based on SPER scores}
\usage{
plotSDG(
  cell.type,
  gene.id,
  score = NULL,
  thershold = 0.3,
  coord = coordinates_test,
  ST.data = ST_expression,
  CoDa.data = KP_local$Z,
  reference.data = allen_reference,
  SPER.data = rf_pair_cor_list,
  plot.dir = "Plots/Feature Gene/"
)
}
\arguments{
\item{cell.type}{Cell Type Name}

\item{gene.id}{Gene ID}

\item{score}{SPER score shown in the plot title}

\item{thershold}{Threshold for the proportion of given cell type; spots whose value larger than this
threshold will be marked in the spatial map}

\item{coord}{Spatial coordinates for 'gg_gene'}

\item{ST.data}{Spatial transcriptomics data, used for the spatial map of the gene's expression}

\item{CoDa.data}{Cell-type spatial compositional data, used for the spatial map of cell-type
distribution}

\item{reference.data}{scRNA-seq reference; should be a Seurat object}

\item{SPER.data}{SPER list to plot the SPER curve}

\item{plot.dir}{Directory to save the plots}
}
\value{
Null
}
\description{
Generate a series of plots of SDGs discovered by 'findSDG': the SPER curve is plotted, and a spatial map
showing the spatial of both the gene and the cell type is generated.
}
