% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotSPER}
\alias{plotSPER}
\title{Visualize SPER scores of a gene/cell-type pair}
\usage{
plotSPER(
  cell.type,
  gene.id,
  SPER.score = NULL,
  cell.thershold = 0.3,
  min.expr = 1,
  coord,
  ST.data,
  CoDa.data,
  reference.data,
  SPER.data,
  dist.list,
  plot.dir = "./Plots/",
  save.plots = F
)
}
\arguments{
\item{cell.type}{Cell Type Name}

\item{gene.id}{Gene ID}

\item{SPER.score}{SPER score shown in the plot title}

\item{cell.thershold}{Threshold for the proportion of given cell type; spots whose value larger than this
threshold will be marked in the spatial map}

\item{min.expr}{Minimal expression value to plot (integer); only spots with given gene's expression greater than this value will be plotted}

\item{coord}{Spatial coordinates for 'gg_gene'}

\item{ST.data}{Spatial transcriptomics data, used for the spatial map of the gene's expression}

\item{CoDa.data}{Cell-type spatial compositional data, used for the spatial map of cell-type
distribution}

\item{reference.data}{scRNA-seq reference; should be a Seurat object}

\item{SPER.data}{SPER list to plot the SPER curve}

\item{dist.list}{A vector containing the list of desired distance values}

\item{plot.dir}{Directory to save the plots}

\item{save.plots}{Whether save plots as files or return them as objects; default is FALSE, which returns plots as objects}
}
\value{
Null
}
\description{
Generate a series of plots of SDGs discovered by 'findSDG': 1) the SPER curve of the indicated pair; 2) the expression level of the gene in scRNA-seq
reference data; 3) a spatial map showing the spatial of both the gene and the cell type is also generated.
}
