% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{probalized}
\alias{probalized}
\title{Normalize a vector: L1 norm}
\usage{
probalized(x)
}
\arguments{
\item{x}{An Input vector}
}
\value{
Normalized vector
}
\description{
This function normalizes a given vector so that the sum of every elements is 1
}
