% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{range_1}
\alias{range_1}
\title{Scale the range of a list of numbers to \link{0, 1}}
\usage{
range_1(x)
}
\arguments{
\item{x}{An Input vector}
}
\value{
Scaled vector
}
\description{
Return the scaled vector whose range is \link{0, 1}
}
