% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatial_comp_data}
\alias{spatial_comp_data}
\title{spatial_comp_data}
\format{
An object of class \code{data.frame} with 2695 rows and 10 columns.
}
\usage{
spatial_comp_data
}
\description{
The cell-type spatial compositional data; a spot-by-cell-type data frame (2695 x 10)
}
\author{
Tianxiao Zhao \email{Tianxiao.Zhao@nyulangone.org}
}
\keyword{Voronoi}
\keyword{connections}
