% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPER.R
\name{SPER}
\alias{SPER}
\title{Calculate the SPER score list}
\usage{
SPER(dist.mat, dist.list, ST.mat, CoDa.data)
}
\arguments{
\item{dist.mat}{Distance matrix after rounded}

\item{dist.list}{Distance list which includes the value of interest in the distance matrix}

\item{ST.mat}{The ST matrix}

\item{CoDa.data}{Cell-type spatial compositional data}
}
\value{
A list of matrix, each representing the SPER subscores at a given distance
}
\description{
This function calculates the SPER score list with the input of the distance matrix, the ST data and
the cell-type spatial CoDa.
}
