% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ST_coordinates}
\alias{ST_coordinates}
\title{ST_coordinates}
\format{
An object of class \code{data.frame} with 2695 rows and 2 columns.
}
\usage{
ST_coordinates
}
\description{
The spatial coordinates of ST spots; a data frame
}
\author{
Tianxiao Zhao \email{Tianxiao.Zhao@nyulangone.org}
}
\keyword{coordinates}
\keyword{of}
\keyword{spots}
