% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ST_expression_data}
\alias{ST_expression_data}
\title{ST_expression_data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2695 rows and 37 columns.
}
\usage{
ST_expression_data
}
\description{
The Spatial Transcriptomics data; a spot-by-gene data frame (2695 x 37)
}
\author{
Tianxiao Zhao \email{Tianxiao.Zhao@nyulangone.org}
}
\keyword{Spatial}
\keyword{Transcriptomics}
\keyword{data}
