% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{visualizePairCor}
\alias{visualizePairCor}
\title{Visualize SPER scores of a gene/cell-type pair}
\usage{
visualizePairCor(
  gene.id,
  cell.type,
  plot.dir,
  dist.list = dist_list,
  reference = allen_reference
)
}
\arguments{
\item{gene.id}{Gene ID}

\item{cell.type}{Name of cell type}

\item{plot.dir}{Directory to save the plots}

\item{dist.list}{Distance list when plotting the SPER curve}

\item{reference}{scRNA-seq reference; should be a Seurat object}
}
\value{
Null
}
\description{
Draw the plot of: 1) the SPER curve of the indicated pair; 2) the expression level of the gene in scRNA-seq
reference data; 3) spatial map using gg_gene
}
