% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPER.R
\name{findSDG}
\alias{findSDG}
\title{Find the spatial dependent gene (SDG) based on SPER scores}
\usage{
findSDG(
  cell.type,
  score.mat,
  extra.gene.list,
  expr.mat,
  LRP.data,
  marker.gene,
  ratio.threshold = 1.5,
  expr.fraction = 0.15,
  method.name = "Multiply",
  only.extra = T,
  plot.feature = T,
  return.numbers = F,
  calc.receptor.frac = F,
  except.marker = T
)
}
\arguments{
\item{cell.type}{Cell Type Name}

\item{score.mat}{Score matrix (SPER, or correlations)}

\item{extra.gene.list}{The gene list of interest: could be extracellular or ligand gene sets}

\item{expr.mat}{Expression prevalence matrix}

\item{LRP.data}{Ligand-receptor pair data}

\item{marker.gene}{The marker gene matrix with a column 'Type' indicating the cell type information and 'Gene' indicating the gene ID}

\item{ratio.threshold}{Threshold on the score: pairs whose scores larger than the threshold will be kept}

\item{expr.fraction}{Threshold on expression prevalence: pairs whose prevalence smaller than the threshold will be kept}

\item{method.name}{Name of the method}

\item{only.extra}{Boolean value: only consider the candidate genes}

\item{plot.feature}{Boolean value: plot the SDG using 'plotFeatureGene'}

\item{return.numbers}{Boolean value: return the statistical results of hypergeometric test}

\item{calc.receptor.frac}{Boolean value: calculate the fraction of receptors that expressed in the target cell types. A column of these receptors must be provided in the extra.gene.list as 'gene2'.}

\item{except.marker}{Boolean value: ignore the marker of the cell types}
}
\value{
Scaled vector
}
\description{
This function helps find the SDG based on the given threshold in both SPER scores and
expression prevalence.
}
